move Containers into their own module, make it a base class
