--- bellini/cupboard.py
@@ -1,15 +1,0 @@
-# =============================================================================
-# IMPORTS
-# =============================================================================
-import abc
-import pint
-import numpy as np
-from .groups import Mixture
-ureg = pint.UnitRegistry()
-
-# =============================================================================
-# MODULE CLASSES
-# =============================================================================
-class ConstantDensitySolution(Mixture):
-    """ Solution with constant density. """
-
