+++ bellini/groups.py
@@ -580,62 +580,3 @@ class Solution(Liquid):
         )
 
         return aliquot, source
-
-# =============================================================================
-# Containers
-# =============================================================================
-
-class Container(Group):
-    """ Simple container for a solution """
-    def __init__(self, solution=None, **values):
-        if solution is not None:
-            assert isinstance(solution, Liquid)
-        super().__init__(solution = solution, **values)
-
-    @property
-    def volume(self):
-        if self.solution is not None:
-            return self.solution.volume
-        else:
-            return Quantity(0.0, VOLUME_UNIT)
-
-    def retrieve_aliquot(self, volume):
-        """ Removes an aliquot and returns it """
-        assert self.solution is not None # TODO: check that volume is enough to remove an aliquot
-        aliquot, source = self.solution.aliquot(volume)
-        self.solution = source
-        return aliquot
-
-    def receive_aliquot(self, solution):
-        if self.solution is not None:
-            self.solution = self.solution + solution
-        else:
-            self.solution = solution
-
-    def __repr__(self):
-        return f"Well containing {self.solution}"
-
-    def observe(self, value, key=None):
-        if key:
-            return getattr(self.solution, value)[key]
-        else:
-            return getattr(self.solution, value)
-
-    def copy(self):
-        return NotImplementedError()
-
-
-class WellArray(Container):
-    """ An array of Containers (e.g. well plate). Must contain an array """
-    def __init__(self, solution=None, **values):
-        assert utils.is_arr(solution.volume.magnitude)
-        super().__init__(solution=solution, **values)
-
-    def subset_aliquot(self, idxs, volume):
-        clear = utils.mask(self.solution.volume.magnitude, idxs, invert=True)
-        select = utils.mask(self.solution.volume.magnitude, idxs, invert=False)
-        aliquot, _ = self.solution[idxs].aliquot(volume)
-        print(self.solution.aliquot(volume)[0].__class__)
-        source = self.solution.aliquot(volume)[0] * select
-        source = source + self.solution * clear
-        return aliquot, source
